+++ Editor/DB/DBEditorWindow.cs
@@ -1,4 +1,6 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using OcDialogue.DB;
 using Sirenix.OdinInspector.Editor;
 using UnityEditor;
@@ -11,6 +13,8 @@ namespace OcDialogue.Editor
         List<string> _DBNames;
         OcDB _currentDB;
         IDBEditor _CurrentDBEditor;
+        static string _searchText;
+        static DialogueEditorWindow.SortMethod _sortMethod;
 
         int CurrentSelectedDBIndex
         {
@@ -40,7 +44,7 @@ namespace OcDialogue.Editor
             base.OnEnable();
             UpdateDBEditor();
         }
-
+        
         protected override void OnBeginDrawEditors()
         {
             GUI.color = new Color(1f, 1f, 2f);
@@ -56,9 +60,31 @@ namespace OcDialogue.Editor
         protected override OdinMenuTree BuildMenuTree()
         {
             var tree = new OdinMenuTree();
-            _CurrentDBEditor?.CreateTree(tree);
 
+            _CurrentDBEditor?.CreateTree(tree);
+            tree.FlatMenuTree.RemoveAll(x => !x.Name.Contains(_searchText));
+            switch (_sortMethod)
+            {
+                case DialogueEditorWindow.SortMethod.Index:
+                    break;
+                case DialogueEditorWindow.SortMethod.Name:
+                    tree.FlatMenuTree = tree.FlatMenuTree.OrderBy(x => x.Name).ToList();
+                    break;
+                case DialogueEditorWindow.SortMethod.NameDescending:
+                    tree.FlatMenuTree = tree.FlatMenuTree.OrderByDescending(x => x.Name).ToList();
+                    break;
+                default:
+                    throw new ArgumentOutOfRangeException();
+            }
+            
             return tree;
+        }
+
+        protected override void DrawMenu()
+        {
+            _searchText = EditorGUILayout.TextField("",_searchText);
+            _sortMethod = (DialogueEditorWindow.SortMethod)EditorGUILayout.EnumPopup("", _sortMethod);
+            base.DrawMenu();
         }
 
         void UpdateDBEditor()
@@ -78,5 +104,6 @@ namespace OcDialogue.Editor
                 UnityEditor.Editor.CreateEditor(DBManager.Instance.DBs[__currentSelectedDBIndex]) as IDBEditor;
             _CurrentDBEditor.Window = this;
         }
+
     }
 }