+++ Editor/Dialogue/DialogueEditorWindow.cs
@@ -1,18 +1,13 @@
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
 using OcDialogue.DB;
-using OcUtility;
-using Sirenix.OdinInspector.Editor;
-using Sirenix.Utilities.Editor;
 using UnityEditor;
 using UnityEditor.Experimental.GraphView;
 using UnityEditor.UIElements;
 using UnityEngine;
 using UnityEngine.UIElements;
-using Object = UnityEngine.Object;
 
 namespace OcDialogue.Editor
 {
@@ -115,7 +110,11 @@ namespace OcDialogue.Editor
         
         void OnDisable()
         {
-            if (_instance == this) _instance = null;
+            if (_instance == this)
+            {
+                _instance = null;
+                AssetDatabase.SaveAssets();
+            }
         }