+++ Editor/Dialogue/DialogueGraphView.cs
@@ -1,23 +1,12 @@
 using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.Diagnostics;
-using System.IO;
 using System.Linq;
-using System.Reflection;
-using System.Runtime.Serialization.Formatters.Binary;
-using System.Text;
-using System.Xml.Serialization;
 using OcDialogue.DB;
-using OcUtility;
-using Unity.EditorCoroutines.Editor;
 using UnityEditor;
 using UnityEditor.Experimental.GraphView;
 using UnityEngine;
 using UnityEngine.UIElements;
 using Debug = UnityEngine.Debug;
-using Object = UnityEngine.Object;
-using PopupWindow = UnityEngine.UIElements.PopupWindow;
 
 namespace OcDialogue.Editor
 {