+++ Runtime/DB/Common/OcData.cs
@@ -20,13 +20,14 @@ namespace OcDialogue
         public string TotalAddress => $"{(Parent == null ? "" : Parent.TotalAddress + "/")}{Address}";
         [HideInTables]
         public OcData parent;
-
+        
         public abstract bool IsTrue(string fieldName, CheckFactor.Operator op, object checkValue);
         public abstract object GetValue(string fieldName);
         public abstract string[] GetFieldNames();
         public abstract void SetValue(string fieldName, DataSetter.Operator op, object value);
-        public abstract DataRowType GetValueType(string fieldName);
-
+        public abstract DataRowType? GetValueType(string fieldName);
+        public abstract event Action changed;
+        
         public static bool IsTrue(bool a, CheckFactor.Operator op, bool b)
         {
             var ia = a == true ? 1 : 0;